vendor_management/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=0.5.0,<1.0
python-dotenv>=0.20.0,<1.0
